% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_boxplot.R
\name{dv_boxplot}
\alias{dv_boxplot}
\title{Create boxplot to overlay on densityviolin}
\usage{
dv_boxplot(
  .data,
  group,
  x,
  box_width = 1,
  vp = "dviolin_plot::violin_panel",
  box_col = "black",
  box_fill = "black",
  median_col = "red",
  show_outliers = FALSE,
  coef = 1.5,
  pch = 16
)
}
\arguments{
\item{.data}{Data from to generate boxplot from}

\item{x}{The variable to use as boxplot values.}

\item{vp}{}

\item{y}{Factor variable representing the y coordinate to plot box.}
}
\value{

}
\description{
Create a boxplot from a vector of data to overlay on the densityviolin plot.
}
