% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviolin.R
\name{dv_define_plot}
\alias{dv_define_plot}
\title{Define density violin plot area}
\usage{
dv_define_plot(xlim, ylim)
}
\arguments{
\item{ylim}{}
}
\value{
A \link[grid:viewport]{grid::vpTree} defining the density violin plot.
}
\description{
Define density violin plot area
}
