% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_grand_summary.R
\name{dv_grand_summary}
\alias{dv_grand_summary}
\title{Add grand summary (e.g., grand median) to plot.}
\usage{
dv_grand_summary(
  x,
  gp = grid::gpar(),
  vp = NULL,
  summary_function = median,
  ...
)
}
\value{

}
\description{
Add grand summary (e.g., grand median) to plot.
}
