% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviolin.R
\name{dv_plot_area}
\alias{dv_plot_area}
\title{Title}
\usage{
dv_plot_area(
  ylabels = c(),
  gp = grid::gpar(),
  vp = "dviolin_plot::violin_panel"
)
}
\arguments{
\item{vp}{}
}
\value{

}
\description{
Title
}
\details{
TODO:   # Add in N underline at top (bold): col=3,row=1
}
