% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviolin.R
\name{dv_value_legend}
\alias{dv_value_legend}
\title{Draw value legend of density violin plot}
\usage{
dv_value_legend(
  label = "",
  order = c("increasing", "decreasing"),
  fill = "black",
  vp = "dviolin_plot::gutter_panel"
)
}
\arguments{
\item{vp}{}

\item{g}{An existing grob/gTree}
}
\value{
A \code{\link[grid:grid.grob]{grid::grob()}} representing the graphic.
}
\description{
Creates a \code{\link[grid:grid.grob]{grid::grob()}} representing a triangular graphic
representing the increasing (or decreasing) intensity across the axis.
}
