% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{estimate_density}
\alias{estimate_density}
\title{Calculate density estimate of vector}
\usage{
estimate_density(x, limits, bw = "nrd0", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calculates the density estimate, using \code{\link[=density]{density()}}, of the given set of values.
}
\details{
This function calculate the density estimate of the distribution. It is rather
simplified, in that it expects a min/max range for calculating the estimates. The
\code{\link[stats:density]{stats::density()}} function uses a heuristic for this purpose. See
\code{\link[=compute_density_limits]{compute_density_limits()}} for a tweaked version of this heuristic. The purpose is
to separate the two components, since density estimation can occur on subset of the
input data. In these cases, the estimation of limits (and density evaluation points)
can then differ which would be a problem.

limits. This is a two-value vector that defaults to be just
past range of the data (see \code{\link[stats:density]{stats::density()}} for details.
Setting it to different values can be done - consider
the following example. Assume the data (in \code{x}) ranges from 0.5-1 but
the possible range of the variable can be 0-1. In this case, we may
want a density over the wider (0-1) range. It will be zero (or close)
much beyond the scale 0.5-1, but will look better as it tends toward 0.
}
\note{
The return result of this function is not exactly that of \code{\link[stats:density]{stats::density()}}. It
returns a \code{\link[tibble:tibble]{tibble::tibble()}}. The columns of the return result are x, y and n.
}
