% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{estimate_density_groups}
\alias{estimate_density_groups}
\title{Calculate density estimates across groups}
\usage{
estimate_density_groups(.data, x, by, limits, ...)
}
\arguments{
\item{.data}{A data frame to use for \code{x} and \code{by}.}

\item{x}{A variable to estimate density from.}

\item{by}{A grouping-type variable (e.g., factor) to calculate \code{x} densities by.}

\item{...}{Any further parameters to \code{\link[stats:density]{stats::density()}}.}
}
\value{
A data.frame with three columns: x, y and by.
}
\description{
Estimate densities of \code{x} by each level of \code{by}.
}
\details{
Given a data.frame (\code{.data}) with values in \code{x}, estimate the density of \code{x}
separately for each of the levels of \code{by}.

For instance, if \code{by} is \code{color} with values \code{c("blue","red","green")} then
compute densities for \code{x} separately for blue, red and green.

The return result of this function is a data frame with three columns: x, y and
by. The x,y values are the results of estimating density (see \code{\link[=estimate_density]{estimate_density()}})
and the value of \code{by} is the specific level value from the \code{by} variable.
}
