% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is_density_grouping_equal}
\alias{is_density_grouping_equal}
\title{Determine if density grouping(s) are equal}
\usage{
is_density_grouping_equal(d)
}
\arguments{
\item{d}{A data.frame from \code{\link[=estimate_density_by]{estimate_density_by()}}.}
}
\value{

}
\description{
A density grouping is a series of density estimates based on
another variable. This function tests if the density estimate x coordinates
are equal.
}
\details{
A density estimation technique evaluates the density of a given
distribution at a grid of points. In the case of a density grouping, that
means each density estimation should have been done at the same grid of
points to be directly comparable. While that is not always required, this
function tests whether or not the density groupings are equal and thus
able to be compared.
}
\note{
A density grouping is created by \code{\link[=estimate_density_by]{estimate_density_by()}}.
}
