% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{scale_density}
\alias{scale_density}
\title{Scale density estimates to 0/1 range}
\usage{
scale_density(y)
}
\arguments{
\item{y}{Numeric vector of values to scale}
}
\value{
A numeric vector scaled to 0/1 by the range of \code{y}.
}
\description{
Scale density estimates to 0/1 range
}
